refactor(site): share wallet and delivery packages with durable payment recovery
