+++ apps/push/src/config.ts
@@ -1,13 +1,8 @@
+import { DEFAULT_NOSTR_RELAYS } from "@linky/linkstr";
 import { resolve } from "node:path";
 
 export const CATCH_UP_LOOKBACK_SECONDS = 3 * 24 * 60 * 60;
 export const SEEN_EVENT_RETENTION_MARGIN_MS = 6 * 60 * 60 * 1000;
-
-const DEFAULT_RELAYS = [
-  "wss://relay.damus.io",
-  "wss://nos.lol",
-  "wss://relay.0xchat.com",
-];
 
 export interface PushServiceConfig {
   port: number;
@@ -106,8 +101,8 @@ function readListEnv(
 function readRelayList(env: Record<string, string | undefined>): string[] {
   const relays = readListEnv(
     env,
-    "PUSH_DEFAULT_RELAYS",
-    DEFAULT_RELAYS.join(","),
+    "PUSH_DEFAULT_NOSTR_RELAYS",
+    DEFAULT_NOSTR_RELAYS.join(","),
   );
   return [...new Set(relays.map(normalizeRelayUrl))];
 }
@@ -138,7 +133,7 @@ function normalizeRelayUrl(value: string): string {
     parsed.protocol = "wss:";
   } else if (parsed.protocol !== "ws:" && parsed.protocol !== "wss:") {
     throw new ConfigError(
-      `PUSH_DEFAULT_RELAYS contains unsupported protocol in ${value}`,
+      `PUSH_DEFAULT_NOSTR_RELAYS contains unsupported protocol in ${value}`,
     );
   }
