+++ apps/site/src/cashu/CashuPage.tsx
@@ -1,1863 +1,36 @@
-import type { MintKeyset, Proof } from "@cashu/cashu-ts";
-import type { FormEvent } from "react";
-import { useEffect, useMemo, useRef, useState } from "react";
-import { isRecord } from "../isRecord";
+import { GENERIC_MINT_ICON_DATA_URL, isLightningAddress } from "@linky/linkshu";
 import { SiteFooter } from "../SiteFooter";
 import { SiteHeaderMenu } from "../SiteHeaderMenu";
-import {
-  getInitialSiteDisplayCurrency,
-  getInitialSiteLocale,
-  siteDisplayCurrencies,
-  siteDisplayCurrencyStorageKey,
-  siteLocaleStorageKey,
-  type SiteDisplayCurrency,
-  type SiteLocale,
-} from "../sitePreferences";
-import { forwardCashuTokenPrivately } from "./nostrGiftWrap";
-import {
-  flushPaymentTelemetryQueue,
-  PAYMENT_ANALYTICS_RECIPIENT_NPUB,
-  queuePaymentTelemetry,
-} from "./paymentTelemetry";
-
-interface LocaleCopy {
-  cashuLabel: string;
-  cashuOptionDescription: string;
-  collapseOptionsLabel: string;
-  czechLabel: string;
-  currencyLabel: string;
-  expandOptionsLabel: string;
-  englishLabel: string;
-  germanLabel: string;
-  githubLabel: string;
-  invalidToken: string;
-  linkyPrimaryAction: string;
-  lightningAddressLabel: string;
-  lightningOptionDescription: string;
-  lightningAddressPlaceholder: string;
-  loadingToken: string;
-  noTokenLoaded: string;
-  nostrLabel: string;
-  openInWalletLabel: string;
-  pageTitle: string;
-  payoutIntro: string;
-  privacyLabel: string;
-  redeemButton: string;
-  redeemConfirmed: string;
-  redeemFailed: string;
-  redeemLnurlComment: string;
-  redeemSuccessAddress: string;
-  redeeming: string;
-  showTokenButton: string;
-  spentInfo: string;
-  statusSpent: string;
-  subtitle: string;
-  switchLabel: string;
-  tokenLabel: string;
-  validUnknown: string;
-}
-
-interface TokenSnapshot {
-  amount: number;
-  iconUrl: string;
-  isValid: boolean;
-  mint: string;
-  mintHost: string;
-  totalAmount: number;
-  unit: string;
-}
-
-interface TokenInspectionError {
-  code: "invalid" | "unknown";
-  detail: string | null;
-}
-
-interface RedeemSuccessState {
-  lightningAddress: string;
-}
-
-type RedeemTelemetryPhase = "invoice_fetch" | "melt";
-
-class RedeemError extends Error {
-  phase: RedeemTelemetryPhase;
-
-  constructor(message: string, phase: RedeemTelemetryPhase) {
-    super(message);
-    this.name = "RedeemError";
-    this.phase = phase;
-  }
-}
-
-interface SiteFiatRates {
-  chfPerBtc: number;
-  czkPerBtc: number;
-  eurPerBtc: number;
-  fetchedAtMs: number;
-  usdPerBtc: number;
-}
-
-interface LnurlPayParameters {
-  callback: string;
-  commentAllowed: number;
-  lightningAddress: string;
-  maxSendable: number;
-  minSendable: number;
-}
-
-type FiatDisplayCurrency = Exclude<SiteDisplayCurrency, "btc" | "sat">;
-
-const fiatRatesStorageKey = "linky.fiat_rates.v1";
-const fiatRatesTtlMs = 10 * 60 * 1000;
-const satsPerBtc = 100_000_000;
-const linkyWebAppUrl = "https://app.linky.fit";
-const nativeLaunchFallbackDelayMs = 700;
-const pwaLaunchFallbackDelayMs = 1600;
-
-const fiatDisplay: Record<
-  FiatDisplayCurrency,
-  { label: string; rate: Exclude<keyof SiteFiatRates, "fetchedAtMs"> }
-> = {
-  chf: { label: "CHF", rate: "chfPerBtc" },
-  czk: { label: "Kč", rate: "czkPerBtc" },
-  eur: { label: "EUR", rate: "eurPerBtc" },
-  usd: { label: "USD", rate: "usdPerBtc" },
-};
-
-const copy: Record<SiteLocale, LocaleCopy> = {
-  cs: {
-    cashuLabel: "Cashu",
-    cashuOptionDescription: "Nascanujte kód vaší cashu peněženkou",
-    collapseOptionsLabel: "Skrýt možnosti ↑",
-    czechLabel: "Čeština",
-    currencyLabel: "Jednotky",
-    expandOptionsLabel: "Další možnosti ↓",
-    englishLabel: "English",
-    germanLabel: "Deutsch",
-    githubLabel: "GitHub",
-    invalidToken: "Utraceno",
-    linkyPrimaryAction: "Vyzvednout v Linky",
-    lightningAddressLabel: "Lightning adresa",
-    lightningOptionDescription: "Vyberte prostředky na vaši lightning adresu.",
-    lightningAddressPlaceholder: "jmeno@linky.fit",
-    loadingToken: "Ověřuji token u mintu…",
-    noTokenLoaded:
-      "Vlož token ručně nebo otevři stránku rovnou s tokenem v URL.",
-    nostrLabel: "Nostr profil",
-    openInWalletLabel: "Otevřít v peněžence",
-    pageTitle: "Vytvoř odkaz pro vyzvednutí bitcoinu na lightning adresu",
-    payoutIntro:
-      "Někdo vám posílá bitcoin. Vyzvednout si ho můžete v aplikaci Linky nebo jakékoliv lightning peněžence",
-    privacyLabel: "Ochrana soukromí",
-    redeemButton: "Vybrat na adresu",
-    redeemConfirmed: "Hotovo",
-    redeemFailed: "Vyzvednutí se nepodařilo.",
-    redeemLnurlComment: "Vybráno pomocí Linky",
-    redeemSuccessAddress: "Prostředky vybrány na {address}",
-    redeeming: "Vyzvedávám…",
-    showTokenButton: "Zobrazit token",
-    spentInfo: "Už to někdo vybral.",
-    statusSpent: "Utraceno",
-    subtitle:
-      "Vložte existující cashu token a vytvořte odkaz, který můžete poslat svému známému.",
-    switchLabel: "Jazyk",
-    tokenLabel: "Cashu token",
-    validUnknown: "Nepodařilo se načíst token.",
-  },
-  en: {
-    cashuLabel: "Cashu",
-    cashuOptionDescription: "Scan the code with your Cashu wallet",
-    collapseOptionsLabel: "Hide options ↑",
-    czechLabel: "Čeština",
-    currencyLabel: "Units",
-    expandOptionsLabel: "Show options ↓",
-    englishLabel: "English",
-    germanLabel: "Deutsch",
-    githubLabel: "GitHub",
-    invalidToken: "Spent",
-    linkyPrimaryAction: "Redeem in Linky",
-    lightningAddressLabel: "Lightning address",
-    lightningOptionDescription: "Withdraw the funds to your Lightning address.",
-    lightningAddressPlaceholder: "name@linky.fit",
-    loadingToken: "Checking the token with the mint…",
-    noTokenLoaded:
-      "Paste a token manually or open the page directly with a token in the URL.",
-    nostrLabel: "Nostr profile",
-    openInWalletLabel: "Open in wallet",
-    pageTitle: "Create a link to redeem bitcoin to a lightning address",
-    payoutIntro:
-      "Someone is sending you bitcoin. You can redeem it in the Linky app or in any Lightning wallet.",
-    privacyLabel: "Privacy Policy",
-    redeemButton: "Redeem to address",
-    redeemConfirmed: "Success",
-    redeemFailed: "Redeem failed.",
-    redeemLnurlComment: "Redeemed with Linky",
-    redeemSuccessAddress: "Funds redeemed to {address}",
-    redeeming: "Redeeming…",
-    showTokenButton: "Show token",
-    spentInfo: "Someone already redeemed it.",
-    statusSpent: "Spent",
-    subtitle:
-      "Paste an existing Cashu token and create a link to redeem bitcoin to a lightning address.",
-    switchLabel: "Language",
-    tokenLabel: "Cashu token",
-    validUnknown: "Could not load the token.",
-  },
-  de: {
-    cashuLabel: "Cashu",
-    cashuOptionDescription: "Scanne den Code mit deiner Cashu-Wallet",
-    collapseOptionsLabel: "Optionen ausblenden ↑",
-    czechLabel: "Čeština",
-    currencyLabel: "Einheiten",
-    expandOptionsLabel: "Weitere Optionen ↓",
-    englishLabel: "English",
-    germanLabel: "Deutsch",
-    githubLabel: "GitHub",
-    invalidToken: "Ausgegeben",
-    linkyPrimaryAction: "In Linky einlösen",
-    lightningAddressLabel: "Lightning-Adresse",
-    lightningOptionDescription:
-      "Lasse das Guthaben an deine Lightning-Adresse auszahlen.",
-    lightningAddressPlaceholder: "name@linky.fit",
-    loadingToken: "Token wird beim Mint geprüft…",
-    noTokenLoaded:
-      "Füge einen Token ein oder öffne die Seite direkt mit einem Token in der URL.",
-    nostrLabel: "Nostr-Profil",
-    openInWalletLabel: "In Wallet öffnen",
-    pageTitle:
-      "Erstelle einen Link zum Einlösen von Bitcoin an eine Lightning-Adresse",
-    payoutIntro:
-      "Jemand sendet dir Bitcoin. Du kannst sie in der Linky-App oder jeder Lightning-Wallet einlösen.",
-    privacyLabel: "Datenschutz",
-    redeemButton: "An Adresse auszahlen",
-    redeemConfirmed: "Erledigt",
-    redeemFailed: "Einlösen fehlgeschlagen.",
-    redeemLnurlComment: "Mit Linky eingelöst",
-    redeemSuccessAddress: "Guthaben an {address} ausgezahlt",
-    redeeming: "Wird eingelöst…",
-    showTokenButton: "Token anzeigen",
-    spentInfo: "Jemand hat ihn bereits eingelöst.",
-    statusSpent: "Ausgegeben",
-    subtitle:
-      "Füge einen vorhandenen Cashu-Token ein und erstelle einen Link, über den Bitcoin an eine Lightning-Adresse ausgezahlt werden können.",
-    switchLabel: "Sprache",
-    tokenLabel: "Cashu-Token",
-    validUnknown: "Der Token konnte nicht geladen werden.",
-  },
-};
-
-const GENERIC_MINT_ICON_SVG =
-  "<svg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 64 64'><defs><linearGradient id='g' x1='0' y1='0' x2='1' y2='1'><stop offset='0%' stop-color='%2314b8a6'/><stop offset='100%' stop-color='%230ea5e9'/></linearGradient></defs><rect width='64' height='64' rx='32' fill='url(%23g)'/><path d='M21 44V20h6l5 10 5-10h6v24h-5V29l-4.5 8h-3L26 29v15z' fill='white'/></svg>";
-
-const GENERIC_MINT_ICON_DATA_URL = `data:image/svg+xml,${GENERIC_MINT_ICON_SVG}`;
-
-const cashuLibPromise = import("@cashu/cashu-ts");
-
-type CashuLib = Awaited<typeof cashuLibPromise>;
-type CashuWalletLike = InstanceType<CashuLib["Wallet"]>;
-type ProofStatesResponse = Awaited<
-  ReturnType<CashuWalletLike["checkProofsStates"]>
->;
-type CashuProofStateLike = ProofStatesResponse[number];
-
-const getErrorMessage = (value: unknown, fallback: string): string => {
-  if (typeof value === "string" && value) return value;
-  if (value instanceof Error) return value.message || fallback;
-
-  if (
-    typeof value === "object" &&
-    value !== null &&
-    "message" in value &&
-    typeof value.message === "string" &&
-    value.message
-  ) {
-    return value.message;
-  }
-
-  return fallback;
-};
-
-const readObjectField = (value: unknown, field: string): unknown => {
-  return isRecord(value) ? value[field] : undefined;
-};
-
-const readStringField = (
-  record: Record<string, unknown>,
-  field: string,
-): string => {
-  const value = record[field];
-  return typeof value === "string" ? value.trim() : "";
-};
-
-const isPositiveNumber = (value: unknown): value is number => {
-  return typeof value === "number" && Number.isFinite(value) && value > 0;
-};
-
-const siteFiatRateKeys: readonly (keyof SiteFiatRates)[] = [
-  "chfPerBtc",
-  "czkPerBtc",
-  "eurPerBtc",
-  "fetchedAtMs",
-  "usdPerBtc",
-];
-
-const isSiteFiatRates = (value: unknown): value is SiteFiatRates => {
-  return (
-    isRecord(value) &&
-    siteFiatRateKeys.every((key) => isPositiveNumber(value[key]))
-  );
-};
-
-const readStoredSiteFiatRates = (): SiteFiatRates | null => {
-  if (typeof window === "undefined") return null;
-
-  try {
-    const raw = window.localStorage.getItem(fiatRatesStorageKey);
-    if (!raw) return null;
-    const parsed: unknown = JSON.parse(raw);
-    return isSiteFiatRates(parsed) ? parsed : null;
-  } catch {
-    return null;
-  }
-};
-
-const storeSiteFiatRates = (value: SiteFiatRates): void => {
-  if (typeof window === "undefined") return;
-
-  try {
-    window.localStorage.setItem(fiatRatesStorageKey, JSON.stringify(value));
-  } catch {
-    // Keep going without cached rates.
-  }
-};
-
-const areSiteFiatRatesStale = (value: SiteFiatRates | null): boolean => {
-  if (!value) return true;
-  return Date.now() - value.fetchedAtMs >= fiatRatesTtlMs;
-};
-
-const readCoinbaseRate = (rates: unknown, currency: string): number | null => {
-  const raw = readObjectField(rates, currency);
-  const parsed = typeof raw === "string" ? Number.parseFloat(raw) : raw;
-  return isPositiveNumber(parsed) ? parsed : null;
-};
-
-const parseFetchedSiteFiatRates = (value: unknown): SiteFiatRates | null => {
-  const rates = readObjectField(readObjectField(value, "data"), "rates");
-  const chfPerBtc = readCoinbaseRate(rates, "CHF");
-  const czkPerBtc = readCoinbaseRate(rates, "CZK");
-  const eurPerBtc = readCoinbaseRate(rates, "EUR");
-  const usdPerBtc = readCoinbaseRate(rates, "USD");
-
-  if (!chfPerBtc || !czkPerBtc || !eurPerBtc || !usdPerBtc) {
-    return null;
-  }
-
-  return {
-    chfPerBtc,
-    czkPerBtc,
-    eurPerBtc,
-    fetchedAtMs: Date.now(),
-    usdPerBtc,
-  };
-};
-
-const fetchSiteFiatRates = async (
-  signal: AbortSignal,
-): Promise<SiteFiatRates | null> => {
-  const url = new URL("https://api.coinbase.com/v2/exchange-rates");
-  url.searchParams.set("currency", "BTC");
-
-  const response = await fetch(url.toString(), {
-    headers: {
-      Accept: "application/json",
-    },
-    signal,
-  });
-
-  if (!response.ok) return null;
-  const payload: unknown = await response.json();
-  return parseFetchedSiteFiatRates(payload);
-};
-
-const normalizeLocale = (lang: SiteLocale): string => {
-  if (lang === "cs") return "cs-CZ";
-  if (lang === "de") return "de-DE";
-  return "en-US";
-};
-
-const formatInteger = (value: number, lang: SiteLocale): string => {
-  return new Intl.NumberFormat(normalizeLocale(lang)).format(
-    Number.isFinite(value) ? Math.trunc(value) : 0,
-  );
-};
-
-const formatCashuDisplayAmount = (
-  amountSat: number,
-  displayCurrency: SiteDisplayCurrency,
-  fiatRates: SiteFiatRates | null,
-  lang: SiteLocale,
-): string => {
-  const normalizedAmount = Number.isFinite(amountSat)
-    ? Math.max(0, Math.trunc(amountSat))
-    : 0;
-
-  if (displayCurrency === "btc") {
-    return `${formatInteger(normalizedAmount, lang)} ₿`;
-  }
-
-  if (displayCurrency !== "sat" && fiatRates) {
-    const { label, rate } = fiatDisplay[displayCurrency];
-    const fiatValue = Math.round(
-      (normalizedAmount / satsPerBtc) * fiatRates[rate],
-    );
-    const approxPrefix = normalizedAmount > 0 ? "~" : "";
-    return `${approxPrefix}${formatInteger(fiatValue, lang)} ${label}`;
-  }
-
-  return `${formatInteger(normalizedAmount, lang)} sat`;
-};
-
-const copyTextToClipboard = async (value: string): Promise<boolean> => {
-  const trimmed = value.trim();
-  if (!trimmed) return false;
-
-  if (navigator.clipboard?.writeText) {
-    try {
-      await navigator.clipboard.writeText(trimmed);
-      return true;
-    } catch {
-      // Fall through to the textarea fallback below.
-    }
-  }
-
-  if (typeof document === "undefined") return false;
-
-  const textarea = document.createElement("textarea");
-  textarea.value = trimmed;
-  textarea.setAttribute("readonly", "true");
-  textarea.style.position = "absolute";
-  textarea.style.opacity = "0";
-  textarea.style.pointerEvents = "none";
-  textarea.style.left = "-9999px";
-  document.body.append(textarea);
-  textarea.select();
-  textarea.setSelectionRange(0, textarea.value.length);
-
-  try {
-    return document.execCommand("copy");
-  } catch {
-    return false;
-  } finally {
-    textarea.remove();
-  }
-};
-
-interface LinkyWalletImportTargets {
-  browserUrl: string;
-  nativeUrl: string;
-  pwaUrl: string;
-}
-
-const buildLinkyWalletImportTargets = (
-  token: string,
-): LinkyWalletImportTargets | null => {
-  const trimmed = token.trim();
-  if (!trimmed) return null;
-  const encodedToken = encodeURIComponent(trimmed);
-  return {
-    browserUrl: `${linkyWebAppUrl}/#wallet?cashu=${encodedToken}`,
-    nativeUrl: `cashu://receive?token=${encodedToken}`,
-    pwaUrl: `web+cashu://receive?token=${encodedToken}`,
-  };
-};
-
-const openLinkyWalletImport = (token: string): void => {
-  const targets = buildLinkyWalletImportTargets(token);
-  if (!targets || typeof window === "undefined") return;
-
-  let finished = false;
-  const cleanupCallbacks: Array<() => void> = [];
-  const timeoutIds: number[] = [];
-
-  const cleanup = () => {
-    while (timeoutIds.length > 0) {
-      const timeoutId = timeoutIds.pop();
-      if (timeoutId !== undefined) {
-        window.clearTimeout(timeoutId);
-      }
-    }
-
-    while (cleanupCallbacks.length > 0) {
-      const callback = cleanupCallbacks.pop();
-      callback?.();
-    }
-  };
-
-  const finish = () => {
-    if (finished) return;
-    finished = true;
-    cleanup();
-  };
-
-  const isDocumentVisible = () =>
-    typeof document === "undefined" || document.visibilityState === "visible";
-
-  const tryNavigate = (url: string) => {
-    if (finished || !isDocumentVisible()) {
-      finish();
-      return;
-    }
-
-    window.location.assign(url);
-  };
-
-  if (typeof document !== "undefined") {
-    const handleVisibilityChange = () => {
-      if (document.visibilityState !== "visible") {
-        finish();
-      }
-    };
-
-    document.addEventListener("visibilitychange", handleVisibilityChange);
-    cleanupCallbacks.push(() => {
-      document.removeEventListener("visibilitychange", handleVisibilityChange);
-    });
-  }
-
-  const handlePageHide = () => {
-    finish();
-  };
-
-  window.addEventListener("pagehide", handlePageHide);
-  cleanupCallbacks.push(() => {
-    window.removeEventListener("pagehide", handlePageHide);
-  });
-
-  timeoutIds.push(
-    window.setTimeout(() => {
-      tryNavigate(targets.pwaUrl);
-    }, nativeLaunchFallbackDelayMs),
-  );
-
-  timeoutIds.push(
-    window.setTimeout(() => {
-      finish();
-      window.location.assign(targets.browserUrl);
-    }, pwaLaunchFallbackDelayMs),
-  );
-
-  tryNavigate(targets.nativeUrl);
-};
-
-const normalizeMintUrl = (value: string): string => {
-  return value.trim().replace(/\/+$/, "");
-};
-
-const isHexString = (value: string): boolean => {
-  return /^[0-9a-f]+$/i.test(value);
-};
-
-const getMintOriginAndHost = (
-  mintUrl: string,
-): { host: string | null; origin: string | null } => {
-  const raw = mintUrl.trim();
-  if (!raw) return { origin: null, host: null };
-
-  try {
-    const url = new URL(raw);
-    return { origin: url.origin, host: url.host };
-  } catch {
-    return { origin: null, host: raw };
-  }
-};
-
-const getMintIconOverride = (host: string | null): string | null => {
-  if (!host) return null;
-  const key = host.toLowerCase();
-  if (key === "cashu.cz") return "https://cashu.cz/icon.webp";
-  if (key === "testnut.cashu.space") {
-    return "https://image.nostr.build/46ee47763c345d2cfa3317f042d332003f498ee281fb42808d47a7d3b9585911.png";
-  }
-  if (key === "mint.minibits.cash") {
-    return "https://play-lh.googleusercontent.com/raLGxOOzbxOsEx25gr-rISzJOdbgVPG11JHuI2yV57TxqPD_fYBof9TRh-vUE-XyhgmN=w40-h480-rw";
-  }
-  if (key === "kashu.me") {
-    return "https://image.nostr.build/ca72a338d053ffa0f283a1399ebc772bef43814e4998c1fff8aa143b1ea6f29e.jpg";
-  }
-  if (key === "cashu.21m.lol") {
-    return "https://em-content.zobj.net/source/apple/391/zany-face_1f92a.png";
-  }
-  return null;
-};
-
-const mintInfoIconKeys = [
-  "icon_url",
-  "iconUrl",
-  "icon",
-  "logo",
-  "image",
-  "image_url",
-  "imageUrl",
-];
-
-const findMintInfoIconValue = (
-  value: unknown,
-  seen: Set<object>,
-): string | null => {
-  if (Array.isArray(value)) {
-    if (seen.has(value)) return null;
-    seen.add(value);
-    for (const inner of value) {
-      const found = findMintInfoIconValue(inner, seen);
-      if (found) return found;
-    }
-    return null;
-  }
-
-  if (!isRecord(value)) return null;
-  if (seen.has(value)) return null;
-  seen.add(value);
-
-  for (const key of mintInfoIconKeys) {
-    const raw = value[key];
-    if (typeof raw !== "string") continue;
-    const trimmed = raw.trim();
-    if (trimmed) return trimmed;
-  }
-
-  for (const inner of Object.values(value)) {
-    const found = findMintInfoIconValue(inner, seen);
-    if (found) return found;
-  }
-
-  return null;
-};
-
-const resolveMintIconUrl = (mintUrl: string, mintInfo: unknown): string => {
-  const { origin, host } = getMintOriginAndHost(mintUrl);
-  const rawIcon = findMintInfoIconValue(mintInfo, new Set());
-  if (rawIcon) {
-    try {
-      return new URL(rawIcon, origin ?? undefined).toString();
-    } catch {
-      return GENERIC_MINT_ICON_DATA_URL;
-    }
-  }
-
-  const override = getMintIconOverride(host);
-  if (override) return override;
-
-  if (origin) {
-    return `${origin}/favicon.ico`;
-  }
-
-  return GENERIC_MINT_ICON_DATA_URL;
-};
-
-const hasToNumber = (value: unknown): value is { toNumber: () => number } => {
-  return (
-    typeof value === "object" &&
-    value !== null &&
-    "toNumber" in value &&
-    typeof value.toNumber === "function"
-  );
-};
-
-const cashuAmountToNumber = (amount: unknown): number => {
-  if (amount === null || amount === undefined) return 0;
-  if (typeof amount === "number") {
-    if (!Number.isFinite(amount)) return 0;
-    return Math.trunc(amount);
-  }
-  if (typeof amount === "bigint") {
-    const next = Number(amount);
-    if (!Number.isSafeInteger(next)) {
-      throw new Error("Cashu amount exceeds safe integer range");
-    }
-    return next;
-  }
-  if (typeof amount === "string") {
-    const next = Number(amount);
-    if (!Number.isFinite(next)) return 0;
-    if (!Number.isSafeInteger(next)) {
-      throw new Error("Cashu amount exceeds safe integer range");
-    }
-    return Math.trunc(next);
-  }
-  if (hasToNumber(amount)) return amount.toNumber();
-  return 0;
-};
-
-const sumProofAmounts = (proofs: readonly Proof[]): number => {
-  let sum = 0;
-
-  for (const proof of proofs) {
-    const amount = cashuAmountToNumber(proof.amount);
-    if (Number.isFinite(amount) && amount > 0) {
-      sum += Math.trunc(amount);
-    }
-  }
-
-  return sum;
-};
-
-const getMeltSwapTargetAmount = (
-  quotedTotalAmount: number,
-  availableAmount: number,
-): number => {
-  const normalizedQuotedTotalAmount = Math.trunc(quotedTotalAmount);
-  const normalizedAvailableAmount = Math.trunc(availableAmount);
-
-  if (
-    !Number.isFinite(normalizedQuotedTotalAmount) ||
-    normalizedQuotedTotalAmount <= 0
-  ) {
-    return 0;
-  }
-
-  if (
-    !Number.isFinite(normalizedAvailableAmount) ||
-    normalizedAvailableAmount <= normalizedQuotedTotalAmount
-  ) {
-    return normalizedQuotedTotalAmount;
-  }
-
-  return normalizedQuotedTotalAmount + 1;
-};
-
-const FULL_BALANCE_PAYMENT_RETRY_FEE_RESERVES = [0, 1, 2, 3, 5, 8, 13, 21];
-
-const getPositiveInteger = (value: string | undefined): number | null => {
-  const parsed = Number(value);
-  if (!Number.isFinite(parsed) || parsed <= 0) return null;
-  return Math.trunc(parsed);
-};
-
-const getPaymentAmountShortage = (errorMessage: string): number | null => {
-  const normalizedMessage = errorMessage.trim();
-
-  const providedNeededMatch = normalizedMessage.match(
-    /provided:\s*(\d+)\s*,\s*needed:\s*(\d+)/i,
-  );
-  if (providedNeededMatch) {
-    const provided = getPositiveInteger(providedNeededMatch[1]);
-    const needed = getPositiveInteger(providedNeededMatch[2]);
-    if (provided !== null && needed !== null && needed > provided) {
-      return needed - provided;
-    }
-  }
-
-  const needHaveMatch = normalizedMessage.match(
-    /need\s*(\d+)\s*,\s*have\s*(\d+)/i,
-  );
-  if (needHaveMatch) {
-    const needed = getPositiveInteger(needHaveMatch[1]);
-    const have = getPositiveInteger(needHaveMatch[2]);
-    if (needed !== null && have !== null && needed > have) {
-      return needed - have;
-    }
-  }
-
-  const feeMatch = normalizedMessage.match(/fee\s*:\s*(\d+)/i);
-  if (feeMatch) {
-    return getPositiveInteger(feeMatch[1]);
-  }
-
-  return null;
-};
-
-const isRetryablePaymentAmountFailure = (errorMessage: string): boolean => {
-  const normalizedMessage = errorMessage.trim().toLowerCase();
-  return (
-    normalizedMessage.includes("insufficient funds") ||
-    normalizedMessage.includes("not enough funds") ||
-    normalizedMessage.includes("not enough balance") ||
-    normalizedMessage.includes("amount out of lnurl range") ||
-    normalizedMessage.includes("not enough inputs provided for melt")
-  );
-};
-
-const buildPaymentAmountAttempts = (
-  requestedAmountSat: number,
-  availableBalanceSat: number,
-): number[] => {
-  const normalizedRequestedAmount = Math.trunc(requestedAmountSat);
-  const normalizedAvailableBalance = Math.trunc(availableBalanceSat);
-
-  if (
-    !Number.isFinite(normalizedRequestedAmount) ||
-    normalizedRequestedAmount <= 0
-  ) {
-    return [];
-  }
-
-  if (normalizedRequestedAmount !== normalizedAvailableBalance) {
-    return [normalizedRequestedAmount];
-  }
-
-  const attempts: number[] = [];
-  for (const feeReserveSat of FULL_BALANCE_PAYMENT_RETRY_FEE_RESERVES) {
-    const candidateAmount = normalizedRequestedAmount - feeReserveSat;
-    if (candidateAmount <= 0 || attempts.includes(candidateAmount)) continue;
-    attempts.push(candidateAmount);
-  }
-
-  return attempts;
-};
-
-const buildPaymentFailureAmountAttempts = (
-  requestedAmountSat: number,
-  errorMessage: string,
-): number[] => {
-  const normalizedRequestedAmount = Math.trunc(requestedAmountSat);
-  if (
-    !Number.isFinite(normalizedRequestedAmount) ||
-    normalizedRequestedAmount <= 1 ||
-    !isRetryablePaymentAmountFailure(errorMessage)
-  ) {
-    return [];
-  }
-
-  const attempts: number[] = [];
-  const pushAttempt = (candidateAmountSat: number) => {
-    const normalizedCandidate = Math.trunc(candidateAmountSat);
-    if (
-      !Number.isFinite(normalizedCandidate) ||
-      normalizedCandidate <= 0 ||
-      normalizedCandidate >= normalizedRequestedAmount ||
-      attempts.includes(normalizedCandidate)
-    ) {
-      return;
-    }
-    attempts.push(normalizedCandidate);
-  };
-
-  const shortage = getPaymentAmountShortage(errorMessage);
-  if (shortage !== null) {
-    pushAttempt(normalizedRequestedAmount - shortage);
-    pushAttempt(normalizedRequestedAmount - shortage - 1);
-  }
-
-  for (const feeReserveSat of FULL_BALANCE_PAYMENT_RETRY_FEE_RESERVES) {
-    if (feeReserveSat <= 0) continue;
-    pushAttempt(normalizedRequestedAmount - feeReserveSat);
-  }
-
-  return attempts;
-};
-
-const getMeltQuoteState = (value: unknown): string => {
-  const quote = readObjectField(value, "quote");
-  return String(readObjectField(quote, "state") ?? "")
-    .trim()
-    .toUpperCase();
-};
-
-const getMeltQuoteStateErrorMessage = (state: string): string => {
-  switch (state) {
-    case "PENDING":
-      return "Lightning payment is still pending at the mint";
-    case "UNPAID":
-      return "Lightning payment was not paid by the mint";
-    default:
-      return state
-        ? `Unexpected melt quote state: ${state}`
-        : "Melt quote was not paid";
-  }
-};
-
-const isFakeLightningMint = (mintUrl: string, mintInfo: unknown): boolean => {
-  const normalizedMintUrl = normalizeMintUrl(mintUrl).toLowerCase();
-  if (normalizedMintUrl.includes("testnut.cashu.space")) {
-    return true;
-  }
-
-  try {
-    const serializedMintInfo = JSON.stringify(mintInfo).toLowerCase();
-    return (
-      serializedMintInfo.includes("fakewallet") ||
-      serializedMintInfo.includes(
-        "all your lightning invoices will always be marked paid",
-      )
-    );
-  } catch {
-    return false;
-  }
-};
-
-const findBestRedeemQuote = async (
-  wallet: CashuWalletLike,
-  lightningAddress: string,
-  requestedAmountCap: number,
-  availableAmount: number,
-  comment?: string,
-): Promise<{
-  feeReserve: number;
-  quote: Awaited<ReturnType<CashuWalletLike["createMeltQuoteBolt11"]>>;
-  quoteAmount: number;
-  swapTargetAmount: number;
-}> => {
-  let low = 1;
-  let high = Math.min(
-    Math.max(1, Math.trunc(requestedAmountCap)),
-    Math.max(1, Math.trunc(availableAmount)),
-  );
-  let best: {
-    feeReserve: number;
-    quote: Awaited<ReturnType<CashuWalletLike["createMeltQuoteBolt11"]>>;
-    quoteAmount: number;
-    swapTargetAmount: number;
-    total: number;
-  } | null = null;
-  let lastError: unknown = null;
-  const lnurlPayParameters = await fetchLnurlPayParameters(lightningAddress);
-
-  while (low <= high) {
-    const requestedAmount = Math.floor((low + high) / 2);
-
-    try {
-      const invoice = await fetchLnurlInvoice(
-        lnurlPayParameters,
-        requestedAmount,
-        comment,
-      );
-      const quote = await wallet.createMeltQuoteBolt11(invoice);
-      const quoteAmount = cashuAmountToNumber(quote.amount);
-      const feeReserve = cashuAmountToNumber(quote.fee_reserve);
-      const total = quoteAmount + feeReserve;
-      const swapTargetAmount = getMeltSwapTargetAmount(total, availableAmount);
-
-      if (
-        !Number.isFinite(quoteAmount) ||
-        !Number.isFinite(feeReserve) ||
-        !Number.isFinite(total) ||
-        !Number.isFinite(swapTargetAmount) ||
-        quoteAmount <= 0 ||
-        total <= 0
-      ) {
-        throw new Error("Invalid melt quote");
-      }
-
-      if (swapTargetAmount > availableAmount) {
-        high = requestedAmount - 1;
-        continue;
-      }
-
-      best = {
-        feeReserve,
-        quote,
-        quoteAmount,
-        swapTargetAmount,
-        total,
-      };
-
-      if (swapTargetAmount === availableAmount) {
-        break;
-      }
-
-      low = requestedAmount + 1;
-    } catch (error) {
-      lastError = error;
-      high = requestedAmount - 1;
-    }
-  }
-
-  if (!best) {
-    throw lastError ?? new Error("Redeem failed");
-  }
-
-  return {
-    feeReserve: best.feeReserve,
-    quote: best.quote,
-    quoteAmount: best.quoteAmount,
-    swapTargetAmount: best.swapTargetAmount,
-  };
-};
-
-const buildProofKey = (proof: Proof): string => {
-  return [
-    proof.id.trim(),
-    proof.secret.trim(),
-    proof.C.trim(),
-    String(cashuAmountToNumber(proof.amount)),
-  ].join("|");
-};
-
-const dedupeProofs = (proofs: readonly Proof[]): Proof[] => {
-  const seen = new Set<string>();
-  const unique: Proof[] = [];
-
-  for (const proof of proofs) {
-    const key = buildProofKey(proof);
-    if (seen.has(key)) continue;
-    seen.add(key);
-    unique.push(proof);
-  }
-
-  return unique;
-};
-
-const filterUnspentProofs = (
-  proofs: readonly Proof[],
-  states: readonly CashuProofStateLike[],
-): Proof[] => {
-  if (states.length !== proofs.length) return [...proofs];
-
-  return proofs.filter((_, index) => states[index]?.state === "UNSPENT");
-};
-
-const sleep = async (delayMs: number): Promise<void> => {
-  await new Promise((resolve) => {
-    window.setTimeout(resolve, delayMs);
-  });
-};
-
-const waitForConfirmedMeltQuote = async (
-  wallet: CashuWalletLike,
-  quote: Awaited<ReturnType<CashuWalletLike["createMeltQuoteBolt11"]>>,
-  initialState: string,
-): Promise<string> => {
-  let state = initialState;
-
-  if (state === "PAID" || state === "UNPAID") {
-    return state;
-  }
-
-  const maxAttempts = 8;
-  for (let attempt = 0; attempt < maxAttempts; attempt += 1) {
-    await sleep(500);
-
-    const checkedQuote = await wallet.checkMeltQuoteBolt11(quote);
-    state = checkedQuote.state;
-
-    if (state === "PAID" || state === "UNPAID") {
-      return state;
-    }
-  }
-
-  return state;
-};
-
-const parseTokenFromSearch = (search: string): string | null => {
-  if (!search.startsWith("?")) return null;
-
-  const raw = search.slice(1).trim();
-  if (!raw) return null;
-
-  const searchParams = new URLSearchParams(raw);
-  const namedToken =
-    searchParams.get("token") ??
-    searchParams.get("cashu") ??
-    searchParams.get("cashutoken");
-  const namedValue = (namedToken ?? "").trim();
-  if (namedValue.startsWith("cashu")) {
-    return namedValue;
-  }
-
-  if (raw.startsWith("cashu")) {
-    return decodeURIComponent(raw);
-  }
-
-  return null;
-};
-
-const parseTokenFromHash = (hash: string): string | null => {
-  if (!hash.startsWith("#")) return null;
-
-  const raw = hash.slice(1).trim();
-  if (!raw) return null;
-
-  const decoded = decodeURIComponent(raw);
-  if (decoded.startsWith("cashu")) {
-    return decoded;
-  }
-
-  const params = new URLSearchParams(raw);
-  const namedToken =
-    params.get("token") ?? params.get("cashu") ?? params.get("cashutoken");
-  const namedValue = (namedToken ?? "").trim();
-  return namedValue.startsWith("cashu") ? namedValue : null;
-};
-
-const getTokenFromUrl = (): {
-  source: "hash" | "search" | null;
-  token: string;
-} => {
-  const hashToken = parseTokenFromHash(window.location.hash);
-  if (hashToken) return { source: "hash", token: hashToken };
-
-  const searchToken = parseTokenFromSearch(window.location.search);
-  if (searchToken) return { source: "search", token: searchToken };
-
-  return { source: null, token: "" };
-};
-
-const replaceHashToken = (token: string): void => {
-  const url = new URL(window.location.href);
-  url.search = "";
-  url.hash = token ? encodeURIComponent(token) : "";
-  window.history.replaceState(null, "", url.toString());
-};
-
-const fetchJson = async (url: string): Promise<Record<string, unknown>> => {
-  const response = await fetch(url, {
-    headers: {
-      Accept: "application/json",
-    },
-  });
-
-  if (!response.ok) {
-    throw new Error(`HTTP ${response.status}`);
-  }
-
-  const data: unknown = await response.json();
-  if (!isRecord(data)) {
-    throw new Error("Invalid JSON response");
-  }
-
-  return data;
-};
-
-const fetchLnurlJson = async (
-  url: string,
-  proxyQuery: Record<string, string>,
-): Promise<Record<string, unknown>> => {
-  try {
-    return await fetchJson(url);
-  } catch {
-    return await fetchJson(`/api/lnurlp?${new URLSearchParams(proxyQuery)}`);
-  }
-};
-
-const stripLightningPrefix = (value: string): string => {
-  return value.replace(/^lightning:/i, "").trim();
-};
-
-const isLightningAddress = (value: string): boolean => {
-  return /^[^@\s]+@[^@\s]+\.[^@\s]+$/.test(stripLightningPrefix(value));
-};
-
-const getLnurlEndpoint = (lightningAddress: string): string => {
-  const trimmed = stripLightningPrefix(lightningAddress);
-  const atIndex = trimmed.lastIndexOf("@");
-  if (atIndex <= 0 || atIndex === trimmed.length - 1) {
-    throw new Error("Invalid lightning address");
-  }
-
-  const user = trimmed.slice(0, atIndex);
-  const domain = trimmed.slice(atIndex + 1);
-  return `https://${domain}/.well-known/lnurlp/${encodeURIComponent(user)}`;
-};
-
-const fetchLnurlPayParameters = async (
-  lightningAddress: string,
-): Promise<LnurlPayParameters> => {
-  if (!isLightningAddress(lightningAddress)) {
-    throw new Error("Invalid lightning address");
-  }
-
-  const payRequest = await fetchLnurlJson(getLnurlEndpoint(lightningAddress), {
-    address: lightningAddress,
-  });
-
-  if (readStringField(payRequest, "status").toUpperCase() === "ERROR") {
-    throw new Error(readStringField(payRequest, "reason") || "LNURL error");
-  }
-
-  const callback = readStringField(payRequest, "callback");
-  const commentAllowed = Number(payRequest.commentAllowed ?? NaN);
-  const minSendable = Number(payRequest.minSendable ?? NaN);
-  const maxSendable = Number(payRequest.maxSendable ?? NaN);
-
-  if (!callback) {
-    throw new Error("LNURL callback missing");
-  }
-
-  if (!Number.isFinite(minSendable) || !Number.isFinite(maxSendable)) {
-    throw new Error("Invalid LNURL amount range");
-  }
-
-  return {
-    callback,
-    commentAllowed,
-    lightningAddress,
-    maxSendable,
-    minSendable,
-  };
-};
-
-const fetchLnurlInvoice = async (
-  payParameters: LnurlPayParameters,
-  amountSat: number,
-  comment?: string,
-): Promise<string> => {
-  const amountMsat = Math.round(amountSat * 1000);
-  if (
-    amountMsat < payParameters.minSendable ||
-    amountMsat > payParameters.maxSendable
-  ) {
-    throw new Error("Amount out of LNURL range");
-  }
-
-  const fetchInvoiceResponse = (comment: string | null) => {
-    const url = new URL(payParameters.callback);
-    url.searchParams.set("amount", String(amountMsat));
-    if (comment) url.searchParams.set("comment", comment);
-    return fetchLnurlJson(url.toString(), {
-      address: payParameters.lightningAddress,
-      amount: String(amountMsat),
-      ...(comment ? { comment } : {}),
-    });
-  };
-
-  const advertisedCommentLimit =
-    payParameters.commentAllowed > 0
-      ? Math.trunc(payParameters.commentAllowed)
-      : null;
-  const trimmedComment =
-    (comment ?? "").trim().slice(0, advertisedCommentLimit ?? 140) || null;
-  const isBestEffortComment =
-    trimmedComment !== null && advertisedCommentLimit === null;
-  const invoiceResponse = isBestEffortComment
-    ? await fetchInvoiceResponse(trimmedComment).catch(() =>
-        fetchInvoiceResponse(null),
-      )
-    : await fetchInvoiceResponse(trimmedComment);
-
-  if (readStringField(invoiceResponse, "status").toUpperCase() === "ERROR") {
-    throw new Error(
-      readStringField(invoiceResponse, "reason") || "LNURL invoice error",
-    );
-  }
-
-  const invoice =
-    readStringField(invoiceResponse, "pr") ||
-    readStringField(invoiceResponse, "paymentRequest");
-  if (!invoice) {
-    throw new Error("Invoice missing");
-  }
-
-  return invoice;
-};
-
-const pickPreferredMintKeyset = (
-  keysets: readonly MintKeyset[],
-  unit: string,
-): MintKeyset | null => {
-  const matches = keysets
-    .filter((keyset) => {
-      return (
-        keyset.active && keyset.unit.trim() === unit && isHexString(keyset.id)
-      );
-    })
-    .sort((left, right) => {
-      return (left.input_fee_ppk ?? 0) - (right.input_fee_ppk ?? 0);
-    });
-
-  return matches[0] ?? null;
-};
-
-const createLoadedWallet = async (
-  lib: CashuLib,
-  mintUrl: string,
-  unit: string,
-): Promise<CashuWalletLike> => {
-  const mint = new lib.Mint(mintUrl);
-  const wallet = new lib.Wallet(mint, { unit });
-
-  try {
-    await wallet.loadMint();
-    return wallet;
-  } catch (error) {
-    const message = getErrorMessage(error, "").toLowerCase();
-    const keysetVerificationFailed =
-      message.includes("couldn't verify keyset id") ||
-      message.includes("short keyset id v2") ||
-      message.includes("got no keysets to map it to") ||
-      message.includes("couldn't map short keyset id");
-
-    if (!keysetVerificationFailed) {
-      throw error;
-    }
-
-    const [mintInfo, keysetsResponse] = await Promise.all([
-      mint.getInfo(),
-      mint.getKeySets(),
-    ]);
-    const keyset = pickPreferredMintKeyset(keysetsResponse.keysets, unit);
-    if (!keyset?.id) {
-      throw new Error(`No active ${unit} keyset found for ${mintUrl}`);
-    }
-
-    const keysResponse = await mint.getKeys(keyset.id);
-    const keys =
-      keysResponse.keysets.find((candidate) => {
-        return candidate.id === keyset.id && candidate.unit === keyset.unit;
-      }) ?? null;
-
-    if (!keys) {
-      throw new Error(`Mint keys for keyset ${keyset.id} are unavailable`);
-    }
-
-    const cache = {
-      mintUrl,
-      keysets: keysetsResponse.keysets.map((candidate) => {
-        if (candidate.id !== keyset.id) return candidate;
-        return { ...candidate, keys: keys.keys };
-      }),
-    };
-    const fallbackWallet = new lib.Wallet(mint, { unit });
-    fallbackWallet.loadMintFromCache(mintInfo, cache);
-    fallbackWallet.bindKeyset(keyset.id);
-    return fallbackWallet;
-  }
-};
-
-const inspectToken = async (token: string): Promise<TokenSnapshot> => {
-  const lib = await cashuLibPromise;
-  const metadata = lib.getTokenMetadata(token);
-  const mint = normalizeMintUrl(metadata.mint);
-  if (!mint) {
-    throw new Error("Token mint missing");
-  }
-
-  const wallet = await createLoadedWallet(lib, mint, "sat");
-
-  const decoded = wallet.decodeToken(token);
-  const proofs = dedupeProofs(
-    Array.isArray(decoded.proofs) ? decoded.proofs : [],
-  );
-  if (proofs.length === 0) {
-    throw new Error("Token proofs missing");
-  }
-
-  const statesResponse = await wallet.checkProofsStates(proofs);
-  const states = Array.isArray(statesResponse)
-    ? statesResponse
-    : [statesResponse];
-  const unspentProofs = filterUnspentProofs(proofs, states);
-  const mintInfo = wallet.getMintInfo ? await wallet.getMintInfo() : null;
-
-  return {
-    amount: sumProofAmounts(unspentProofs),
-    iconUrl: resolveMintIconUrl(mint, mintInfo),
-    isValid: unspentProofs.length > 0,
-    mint,
-    mintHost: getMintOriginAndHost(mint).host ?? mint,
-    totalAmount: sumProofAmounts(proofs),
-    unit: decoded.unit?.trim() || "sat",
-  };
-};
-
-const redeemToken = async (
-  token: string,
-  lightningAddress: string,
-  comment?: string,
-): Promise<{
-  amountSent: number;
-  changeAmount: number;
-  changeToken: string | null;
-  feePaid: number | null;
-  mint: string;
-}> => {
-  const lib = await cashuLibPromise;
-  const metadata = lib.getTokenMetadata(token);
-  const mint = normalizeMintUrl(metadata.mint);
-  if (!mint) {
-    throw new Error("Token mint missing");
-  }
-
-  const wallet = await createLoadedWallet(lib, mint, "sat");
-  const mintInfo = wallet.getMintInfo ? await wallet.getMintInfo() : null;
-  if (isFakeLightningMint(mint, mintInfo)) {
-    throw new RedeemError(
-      "This testing mint uses fake Lightning settlement, so no real Lightning payment can be completed",
-      "melt",
-    );
-  }
-
-  const decoded = wallet.decodeToken(token);
-  const proofs = dedupeProofs(
-    Array.isArray(decoded.proofs) ? decoded.proofs : [],
-  );
-  const statesResponse = await wallet.checkProofsStates(proofs);
-  const states = Array.isArray(statesResponse)
-    ? statesResponse
-    : [statesResponse];
-  const unspentProofs = filterUnspentProofs(proofs, states);
-  const availableAmount = sumProofAmounts(unspentProofs);
-
-  if (availableAmount <= 0) {
-    throw new Error("Token is already spent");
-  }
-  const queuedAmountAttempts = buildPaymentAmountAttempts(
-    availableAmount,
-    availableAmount,
-  );
-  const seenAmountAttempts = new Set(queuedAmountAttempts);
-  let finalError: unknown = null;
-  let finalPhase: RedeemTelemetryPhase = "melt";
-
-  for (
-    let attemptIndex = 0;
-    attemptIndex < queuedAmountAttempts.length;
-    attemptIndex += 1
-  ) {
-    const attemptedAmount = queuedAmountAttempts[attemptIndex];
-    const hasLowerAmountFallback =
-      attemptIndex < queuedAmountAttempts.length - 1;
-    const queueLowerAmountAttempts = (errorMessage: string): boolean => {
-      const retryAttempts = buildPaymentFailureAmountAttempts(
-        attemptedAmount,
-        errorMessage,
-      );
-      let queuedAny = false;
-      for (const retryAmount of retryAttempts) {
-        if (seenAmountAttempts.has(retryAmount)) continue;
-        seenAmountAttempts.add(retryAmount);
-        queuedAmountAttempts.push(retryAmount);
-        queuedAny = true;
-      }
-      return queuedAny;
-    };
-
-    let bestQuote: Awaited<ReturnType<typeof findBestRedeemQuote>>;
-    try {
-      bestQuote = await findBestRedeemQuote(
-        wallet,
-        lightningAddress,
-        attemptedAmount,
-        availableAmount,
-        comment,
-      );
-    } catch (error) {
-      const errorMessage = getErrorMessage(error, "LNURL invoice error");
-      if (
-        (hasLowerAmountFallback || queueLowerAmountAttempts(errorMessage)) &&
-        isRetryablePaymentAmountFailure(errorMessage)
-      ) {
-        continue;
-      }
-
-      finalError = error;
-      finalPhase = "invoice_fetch";
-      break;
-    }
-
-    let swapped: Awaited<ReturnType<CashuWalletLike["send"]>>;
-    try {
-      swapped = await wallet.send(bestQuote.swapTargetAmount, unspentProofs);
-    } catch (error) {
-      const errorMessage = getErrorMessage(error, "Swap failed");
-      if (
-        (hasLowerAmountFallback || queueLowerAmountAttempts(errorMessage)) &&
-        isRetryablePaymentAmountFailure(errorMessage)
-      ) {
-        continue;
-      }
-
-      finalError = error;
-      finalPhase = "melt";
-      break;
-    }
-
-    let melt: Awaited<ReturnType<CashuWalletLike["meltProofsBolt11"]>>;
-
-    try {
-      melt = await wallet.meltProofsBolt11(bestQuote.quote, swapped.send);
-    } catch (error) {
-      const errorMessage = getErrorMessage(error, "Melt failed");
-      if (
-        (hasLowerAmountFallback || queueLowerAmountAttempts(errorMessage)) &&
-        isRetryablePaymentAmountFailure(errorMessage)
-      ) {
-        continue;
-      }
-
-      finalError = error;
-      finalPhase = "melt";
-      break;
-    }
-
-    let meltQuoteState = getMeltQuoteState(melt);
-    try {
-      meltQuoteState = await waitForConfirmedMeltQuote(
-        wallet,
-        bestQuote.quote,
-        meltQuoteState,
-      );
-    } catch {
-      finalError = new Error(
-        "Could not confirm Lightning payment status with the mint",
-      );
-      finalPhase = "melt";
-      break;
-    }
-
-    if (meltQuoteState !== "PAID") {
-      const error = new Error(getMeltQuoteStateErrorMessage(meltQuoteState));
-      const errorMessage = getErrorMessage(error, "Melt quote was not paid");
-      if (
-        (hasLowerAmountFallback || queueLowerAmountAttempts(errorMessage)) &&
-        isRetryablePaymentAmountFailure(errorMessage)
-      ) {
-        continue;
-      }
-
-      finalError = error;
-      finalPhase = "melt";
-      break;
-    }
-
-    const changeProofs = [
-      ...swapped.keep,
-      ...(Array.isArray(melt.change) ? melt.change : []),
-    ];
-    const feePaidRaw =
-      readObjectField(melt, "fee_paid") ??
-      readObjectField(melt, "feePaid") ??
-      readObjectField(melt, "fee") ??
-      bestQuote.feeReserve;
-    const feePaid = Number(feePaidRaw ?? NaN);
-
-    return {
-      amountSent: bestQuote.quoteAmount,
-      changeAmount: sumProofAmounts(changeProofs),
-      changeToken:
-        changeProofs.length > 0
-          ? lib.getEncodedToken({
-              mint,
-              proofs: changeProofs,
-              unit: decoded.unit?.trim() || "sat",
-            })
-          : null,
-      feePaid:
-        Number.isFinite(feePaid) && feePaid > 0 ? Math.trunc(feePaid) : null,
-      mint,
-    };
-  }
-
-  throw new RedeemError(
-    getErrorMessage(finalError, "Redeem failed"),
-    finalPhase,
-  );
-};
-
+import { useCashuPage } from "./useCashuPage";
 function CashuPage() {
-  const [locale, setLocale] = useState<SiteLocale>(getInitialSiteLocale);
-  const [displayCurrency, setDisplayCurrency] = useState<SiteDisplayCurrency>(
-    getInitialSiteDisplayCurrency,
-  );
-  const [fiatRates, setFiatRates] = useState<SiteFiatRates | null>(() =>
-    readStoredSiteFiatRates(),
-  );
-  const [tokenInput, setTokenInput] = useState("");
-  const [activeToken, setActiveToken] = useState("");
-  const [tokenState, setTokenState] = useState<TokenSnapshot | null>(null);
-  const [tokenError, setTokenError] = useState<TokenInspectionError | null>(
-    null,
-  );
-  const [lightningAddress, setLightningAddress] = useState("");
-  const [redeemError, setRedeemError] = useState<string | null>(null);
-  const [redeemSuccess, setRedeemSuccess] = useState<RedeemSuccessState | null>(
-    null,
-  );
-  const [isInspecting, setIsInspecting] = useState(false);
-  const [isRedeeming, setIsRedeeming] = useState(false);
-  const [isAdditionalOptionsVisible, setIsAdditionalOptionsVisible] =
-    useState(false);
-  const [mintIconSrc, setMintIconSrc] = useState(GENERIC_MINT_ICON_DATA_URL);
-  const [tokenQr, setTokenQr] = useState<string | null>(null);
-  const redeemSubmitLockedRef = useRef(false);
-  const activeCopy = useMemo(() => copy[locale], [locale]);
-  const tokenErrorMessage =
-    tokenError?.detail ??
-    (tokenError?.code === "invalid"
-      ? activeCopy.invalidToken
-      : tokenError?.code === "unknown"
-        ? activeCopy.validUnknown
-        : null);
-  const displayedTokenAmount = tokenState?.isValid
-    ? (tokenState.amount ?? 0)
-    : (tokenState?.totalAmount ?? 0);
-  const displayedTokenAmountText = useMemo(
-    () =>
-      formatCashuDisplayAmount(
-        displayedTokenAmount,
-        displayCurrency,
-        fiatRates,
-        locale,
-      ),
-    [displayCurrency, displayedTokenAmount, fiatRates, locale],
-  );
-  const cycleDisplayCurrency = () => {
-    const currentIndex = siteDisplayCurrencies.indexOf(displayCurrency);
-    const nextIndex = (currentIndex + 1) % siteDisplayCurrencies.length;
-    setDisplayCurrency(siteDisplayCurrencies[nextIndex] ?? "sat");
-  };
-
-  useEffect(() => {
-    document.documentElement.lang = locale;
-  }, [locale]);
-
-  useEffect(() => {
-    window.localStorage.setItem(siteLocaleStorageKey, locale);
-  }, [locale]);
-
-  useEffect(() => {
-    window.localStorage.setItem(siteDisplayCurrencyStorageKey, displayCurrency);
-  }, [displayCurrency]);
-
-  useEffect(() => {
-    let cancelled = false;
-    let activeController: AbortController | null = null;
-
-    const syncRates = async () => {
-      const cached = readStoredSiteFiatRates();
-      if (!cancelled) setFiatRates(cached);
-      if (!areSiteFiatRatesStale(cached)) return;
-
-      const controller = new AbortController();
-      activeController = controller;
-
-      try {
-        const next = await fetchSiteFiatRates(controller.signal);
-        if (!next || cancelled) return;
-        storeSiteFiatRates(next);
-        setFiatRates(next);
-      } catch {
-        // Ignore exchange-rate fetch errors and keep last cached value.
-      } finally {
-        if (activeController === controller) {
-          activeController = null;
-        }
-      }
-    };
-
-    void syncRates();
-    const intervalId = window.setInterval(() => {
-      void syncRates();
-    }, fiatRatesTtlMs);
-
-    return () => {
-      cancelled = true;
-      if (activeController) activeController.abort();
-      window.clearInterval(intervalId);
-    };
-  }, []);
-
-  useEffect(() => {
-    void flushPaymentTelemetryQueue();
-
-    const handleOnline = () => {
-      void flushPaymentTelemetryQueue();
-    };
-
-    window.addEventListener("online", handleOnline);
-    return () => {
-      window.removeEventListener("online", handleOnline);
-    };
-  }, []);
-
-  useEffect(() => {
-    const syncFromUrl = () => {
-      const next = getTokenFromUrl();
-      setTokenInput(next.token);
-      setActiveToken(next.token);
-      setIsAdditionalOptionsVisible(false);
-      setRedeemSuccess(null);
-      setRedeemError(null);
-      setTokenQr(null);
-      redeemSubmitLockedRef.current = false;
-
-      if (next.source === "search" && next.token) {
-        replaceHashToken(next.token);
-      }
-    };
-
-    syncFromUrl();
-    window.addEventListener("hashchange", syncFromUrl);
-
-    return () => {
-      window.removeEventListener("hashchange", syncFromUrl);
-    };
-  }, []);
-
-  useEffect(() => {
-    let cancelled = false;
-
-    const generateQr = async () => {
-      if (!activeToken.trim() || !isAdditionalOptionsVisible) {
-        setTokenQr(null);
-        return;
-      }
-
-      try {
-        const QRCode = await import("qrcode");
-        const qr = await QRCode.toDataURL(activeToken, {
-          errorCorrectionLevel: "M",
-          margin: 1,
-          width: 420,
-        });
-
-        if (!cancelled) {
-          setTokenQr(qr);
-        }
-      } catch {
-        if (!cancelled) {
-          setTokenQr(null);
-        }
-      }
-    };
-
-    void generateQr();
-
-    return () => {
-      cancelled = true;
-    };
-  }, [activeToken, isAdditionalOptionsVisible]);
-
-  useEffect(() => {
-    const trimmedToken = activeToken.trim();
-    if (!trimmedToken) {
-      setTokenState(null);
-      setTokenError(null);
-      setMintIconSrc(GENERIC_MINT_ICON_DATA_URL);
-      return;
-    }
-
-    let cancelled = false;
-
-    setIsInspecting(true);
-    setTokenError(null);
-
-    void inspectToken(trimmedToken)
-      .then((snapshot) => {
-        if (cancelled) return;
-        setTokenState(snapshot);
-        setMintIconSrc(snapshot.iconUrl);
-        if (!snapshot.isValid) {
-          setTokenError({ code: "invalid", detail: null });
-          return;
-        }
-
-        setTokenError(null);
-      })
-      .catch(() => {
-        if (cancelled) return;
-        setTokenState(null);
-        setMintIconSrc(GENERIC_MINT_ICON_DATA_URL);
-        setTokenError({
-          code: "unknown",
-          detail: null,
-        });
-      })
-      .finally(() => {
-        if (cancelled) return;
-        setIsInspecting(false);
-      });
-
-    return () => {
-      cancelled = true;
-    };
-  }, [activeToken]);
-
-  const handleInspectSubmit = (event: FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-    const trimmedToken = tokenInput.trim();
-    replaceHashToken(trimmedToken);
-    setActiveToken(trimmedToken);
-    setRedeemSuccess(null);
-    setRedeemError(null);
-    redeemSubmitLockedRef.current = false;
-  };
-
-  const handleRedeemSubmit = async (event: FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-    if (redeemSubmitLockedRef.current) {
-      return;
-    }
-
-    const trimmedToken = activeToken.trim();
-    const trimmedAddress = stripLightningPrefix(lightningAddress)
-      .trim()
-      .toLowerCase();
-    if (!trimmedToken || !isLightningAddress(trimmedAddress)) return;
-
-    setRedeemError(null);
-    redeemSubmitLockedRef.current = true;
-    setIsRedeeming(true);
-
-    try {
-      const result = await redeemToken(
-        trimmedToken,
-        trimmedAddress,
-        activeCopy.redeemLnurlComment,
-      );
-      const nextToken = result.changeToken?.trim() ?? "";
-
-      if (nextToken && result.changeAmount > 0) {
-        try {
-          await forwardCashuTokenPrivately({
-            recipientNpub: PAYMENT_ANALYTICS_RECIPIENT_NPUB,
-            token: nextToken,
-          });
-        } catch {
-          redeemSubmitLockedRef.current = false;
-          replaceHashToken(nextToken);
-          setTokenInput(nextToken);
-          setActiveToken(nextToken);
-          return;
-        }
-      }
-
-      queuePaymentTelemetry({
-        amount: result.amountSent,
-        direction: "out",
-        fee: result.feePaid,
-        method: "lightning_address",
-        mint: result.mint,
-        phase: "complete",
-        status: "ok",
-      });
-      void flushPaymentTelemetryQueue();
-
-      replaceHashToken("");
-      setTokenInput("");
-      setActiveToken("");
-      setTokenState(null);
-      setRedeemError(null);
-      setRedeemSuccess({
-        lightningAddress: trimmedAddress,
-      });
-    } catch (error) {
-      const phase = error instanceof RedeemError ? error.phase : "melt";
-      const message = getErrorMessage(error, activeCopy.redeemFailed);
-      setRedeemError(message);
-      queuePaymentTelemetry({
-        amount: tokenState?.amount ?? null,
-        direction: "out",
-        error: message,
-        fee: null,
-        method: "lightning_address",
-        mint: tokenState?.mint ?? null,
-        phase,
-        status: "error",
-      });
-      void flushPaymentTelemetryQueue();
-      redeemSubmitLockedRef.current = false;
-    } finally {
-      setIsRedeeming(false);
-    }
-  };
-
-  const handleCopyToken = async () => {
-    await copyTextToClipboard(activeToken);
-  };
-
-  const handleOpenInWallet = () => {
-    openLinkyWalletImport(activeToken);
-  };
+  const {
+    locale,
+    setLocale,
+    tokenInput,
+    setTokenInput,
+    activeToken,
+    tokenState,
+    lightningAddress,
+    setLightningAddress,
+    redeemError,
+    setRedeemError,
+    redeemSuccess,
+    isInspecting,
+    isRedeeming,
+    isAdditionalOptionsVisible,
+    setIsAdditionalOptionsVisible,
+    mintIconSrc,
+    setMintIconSrc,
+    tokenQr,
+    activeCopy,
+    tokenErrorMessage,
+    displayedTokenAmountText,
+    cycleDisplayCurrency,
+    handleInspectSubmit,
+    handleRedeemSubmit,
+    handleCopyToken,
+    handleOpenInWallet,
+  } = useCashuPage();
 
   return (
     <main className="cashu-shell">
@@ -2110,5 +283,4 @@ function CashuPage() {
     </main>
   );
 }
-
 export default CashuPage;
