+++ apps/site/src/main.tsx
@@ -1,7 +1,6 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import App from "./App";
-import "./browserPolyfills";
 import "./index.css";
 
 const container = document.getElementById("root");
