+++ apps/site/tsconfig.node.json
@@ -1,7 +1,14 @@
 {
   "extends": "@linky/config/tsconfig/base",
   "compilerOptions": {
-    "tsBuildInfoFile": "./node_modules/.tmp/tsconfig.node.tsbuildinfo"
+    "tsBuildInfoFile": "./node_modules/.tmp/tsconfig.node.tsbuildinfo",
+    "target": "ES2022"
   },
-  "include": ["vite.config.ts", "vitest.config.ts", "api/**/*.ts"]
+  "include": [
+    "vite.config.ts",
+    "vitest.config.ts",
+    "api/**/*.ts",
+    "playwright.config.ts",
+    "tests/**/*.ts"
+  ]
 }
