+++ apps/web-app/src/app/lib/paymentTelemetry.ts
@@ -1,3 +1,5 @@
+import { classifyPaymentErrorCode } from "@linky/linkstr";
+export { classifyPaymentErrorCode } from "@linky/linkstr";
 import {
   getTelemetryAppHost,
   getTelemetryAppRuntime,
@@ -84,36 +86,6 @@ export const normalizePaymentTelemetryMint = (
   const normalized = normalizeMintUrl(value);
   if (!normalized) return null;
   return normalized.slice(0, 500);
-};
-
-export const classifyPaymentErrorCode = (
-  value: string | null | undefined,
-): string | null => {
-  const text = String(value ?? "")
-    .trim()
-    .toLowerCase();
-  if (!text) return null;
-  if (
-    text.includes("short keyset id v2") ||
-    text.includes("got no keysets to map it to") ||
-    text.includes("couldn't map short keyset id")
-  ) {
-    return "short_keyset_id_unmapped";
-  }
-  if (text.includes("offline")) return "offline";
-  if (text.includes("timeout") || text.includes("timed out")) return "timeout";
-  if (text.includes("insufficient")) return "insufficient";
-  if (text.includes("duplicate")) return "duplicate";
-  if (text.includes("already signed")) return "outputs_already_signed";
-  if (text.includes("publish")) return "publish_failed";
-  if (text.includes("invoice")) return "invoice_failed";
-  if (text.includes("mint")) return "mint_failed";
-  if (text.includes("lnurl")) return "lnurl_failed";
-  if (text.includes("network") || text.includes("fetch")) return "network";
-  if (text.includes("invalid npub")) return "invalid_npub";
-  if (text.includes("invalid nsec")) return "invalid_nsec";
-  if (text.includes("invalid amount")) return "invalid_amount";
-  return "unknown";
 };
 
 const isDeclinedPaymentErrorCode = (value: string | null): boolean => {
