+++ apps/web-app/src/utils/mint.ts
@@ -5,6 +5,12 @@ export interface MintIcon {
   url: string | null;
 }
 
+import { isTestMintUrl as sharedIsTestMintUrl } from "@linky/linkshu";
+import { GENERIC_MINT_ICON_DATA_URL } from "@linky/linkshu";
+export {
+  GENERIC_MINT_ICON_DATA_URL,
+  getMintIconOverride,
+} from "@linky/linkshu";
 const envMainMintUrl = (import.meta.env.VITE_MAIN_MINT_URL ?? "").trim();
 
 export const MAIN_MINT_URL = envMainMintUrl || "https://cashu.cz";
@@ -26,17 +32,10 @@ export const PRESET_MINTS = envMainMintUrl
       ...PRODUCTION_MINTS.slice(1),
     ];
 
-const TEST_MINTS = ["https://testnut.cashu.space"];
-
 export const CASHU_DEFAULT_MINT_OVERRIDE_STORAGE_KEY =
   "linky.cashu.defaultMintOverride.v1";
 
 export const CASHU_SEEN_MINTS_STORAGE_KEY = "linky.cashu.seenMints.v1";
-
-const GENERIC_MINT_ICON_SVG =
-  "<svg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 64 64'><defs><linearGradient id='g' x1='0' y1='0' x2='1' y2='1'><stop offset='0%' stop-color='%2314b8a6'/><stop offset='100%' stop-color='%230ea5e9'/></linearGradient></defs><rect width='64' height='64' rx='32' fill='url(%23g)'/><path d='M21 44V20h6l5 10 5-10h6v24h-5V29l-4.5 8h-3L26 29v15z' fill='white'/></svg>";
-
-export const GENERIC_MINT_ICON_DATA_URL = `data:image/svg+xml,${GENERIC_MINT_ICON_SVG}`;
 
 interface MintStructuredValue {
   toString(): string;
@@ -125,19 +124,8 @@ export const getMintOriginAndHost = (
   }
 };
 
-export const isTestMintUrl = (mint: MintStringInput): boolean => {
-  const normalizedMint = normalizeMintUrl(mint).toLowerCase();
-  if (!normalizedMint) return false;
-
-  // Local dev mints (docker-compose FakeWallet) behave like test mints.
-  const { host } = getMintOriginAndHost(normalizedMint);
-  const hostname = (host ?? "").replace(/:\d+$/, "");
-  if (hostname === "localhost" || hostname === "127.0.0.1") return true;
-
-  return TEST_MINTS.some(
-    (testMint) => normalizeMintUrl(testMint).toLowerCase() === normalizedMint,
-  );
-};
+export const isTestMintUrl = (mint: MintStringInput): boolean =>
+  sharedIsTestMintUrl(normalizeMintUrl(mint));
 
 export const getNextMintIconUrl = (
   currentUrl: string | null,
@@ -182,30 +170,6 @@ export const extractPpk = (value: PpkSearchValue): number | null => {
         queue.push({ value: inner, depth: depth + 1 });
       }
     }
-  }
-  return null;
-};
-
-export const getMintIconOverride = (host: string | null) => {
-  if (!host) return null;
-  const key = host.toLowerCase();
-  if (key === "cashu.cz") {
-    return "https://cashu.cz/icon.webp";
-  }
-  if (key === "testnut.cashu.space") {
-    return "https://image.nostr.build/46ee47763c345d2cfa3317f042d332003f498ee281fb42808d47a7d3b9585911.png";
-  }
-  if (key === "mint.minibits.cash") {
-    return "https://play-lh.googleusercontent.com/raLGxOOzbxOsEx25gr-rISzJOdbgVPG11JHuI2yV57TxqPD_fYBof9TRh-vUE-XyhgmN=w40-h480-rw";
-  }
-  if (key === "linky.cashu.cz") {
-    return "https://linky-weld.vercel.app/icon.svg";
-  }
-  if (key === "kashu.me") {
-    return "https://image.nostr.build/ca72a338d053ffa0f283a1399ebc772bef43814e4998c1fff8aa143b1ea6f29e.jpg";
-  }
-  if (key === "cashu.21m.lol") {
-    return "https://em-content.zobj.net/source/apple/391/zany-face_1f92a.png";
   }
   return null;
 };
