+++ docs/architecture.md
@@ -9,6 +9,9 @@ Architectural decisions and behavioral constraints for Linky. Keep this file up 
 - `apps/site/cashu/` is a separate MPA entry for direct `/cashu` loads; Cashu token privacy should prefer hash fragments (`/cashu/#cashu...`) because query strings are sent to the server on the initial request
 - Landing page is static copy plus product screenshots (`apps/site/public/*.png`); the floating CTA offers Web app, Google Play, Android APK, and Zapstore targets, defaults to Web app, and prefers Google Play when the browser is on an Android phone
 - Landing page and `/cashu` share the top-right language selector (`SiteHeaderMenu`) and the footer (`SiteFooter`); locale and display currency persist under `linky.lang` / `linky.display_currency.v1` (`sitePreferences.ts`). `/cashu` cycles the display currency by tapping the token amount, with `₿` still meaning sats rather than whole bitcoin
+- `/cashu` is a consumer of `@linky/linkshu` and `@linky/linkstr`. `CashuPage` renders the UI, `useCashuPage` owns its state and handlers, and `cashu/wallet.ts` composes package operations. Linkshu owns LNURL parsing and invoice validation, amount fallback, mint loading/proof state, icon metadata/overrides and fiat-rate decoding/fetching. LNURL functions accept an optional HTTP fallback supplied by each app: the web/native adapter retains its hosted proxy, while the site uses the address-only proxy below. Existing web-app LNURL tests exercise that adapter.
+- Site wallets persist seed, token rows, counters and the exact pending melt quote per normalized token under `linky.site.wallet.<hash>`. A Web Lock serializes operations across tabs. Storage must succeed before spending. `MeltDraft.quoteId` reuses the saved quote, verifies its invoice, and `Melt.status` checks settlement without paying again. Reload recovery waits on PENDING, reports PAID to the original address, or restores deterministic outputs and retries only after UNPAID. Change is recovered through `Restore`; reserved rows are released only after a terminal mint answer. Fake mints cannot claim real Lightning delivery; the smoke build explicitly enables `VITE_ALLOW_TEST_MINT=1`.
+- Site telemetry uses linkstr's durable `Outbox` through `runLinkstr`. The old queue key remains a Schema-validated staging buffer for events captured before enqueue; legacy entries migrate there without retaining a second retry engine. A Web Lock serializes workers across tabs; linkstr owns retries and recipient relay acknowledgements. Change forwarding uses `Chat.sendToken` with an ephemeral identity, and failed forwarding keeps a redeemable token visible. Relay defaults, analytics recipient and error classification are shared with the web app.
 - `/cashu` reaches LNURL-pay endpoints directly first and falls back to the same-origin `/api/lnurlp` helper only for CORS-blocked providers; that helper takes a lightning address (plus optional `amount`/`comment`), performs both LNURL hops itself against public https hosts, and returns no CORS headers, so it cannot be used as a general fetch proxy
 - All outbound HTTP from `apps/site/api/` goes through `api/_safeFetch.ts`. Every hop — the LNURL well-known endpoint, the callback it returns, the npub.cash forwards, and any redirect target — must be `https:` on a plain public hostname, is resolved with `dns.lookup({ all: true })` and rejected if any address is loopback, private, link-local, CGNAT, multicast, IPv4-mapped, or otherwise outside public unicast space (`isPublicAddress`), and then connects through an undici `Agent` whose `connect.lookup` returns exactly the validated addresses, so TLS SNI and `Host` keep the hostname while a DNS rebind between check and connect cannot redirect the socket. Redirects are followed manually (at most 3, each re-validated), every hop has a 12 s timeout, and response bodies are capped at 64 KiB
 - `/cashu` token launch now prefers the native `cashu://receive?token=...` deep link, falls back to installed PWA handling via `web+cashu://receive?token=...`, and only then falls back to the browser `https://app.linky.fit/#wallet?cashu=...` import route
