+++ packages/linkshu/package.json
@@ -7,7 +7,8 @@
   "main": "./src/index.ts",
   "types": "./src/index.ts",
   "exports": {
-    ".": "./src/index.ts"
+    ".": "./src/index.ts",
+    "./lightning-address": "./src/lnurl/lightningAddress.ts"
   },
   "scripts": {
     "eslint": "eslint --fix .",
@@ -19,7 +20,8 @@
   "dependencies": {
     "@cashu/cashu-ts": "^4.0.0",
     "effect": "^3.19.19",
-    "@scure/base": "^2.0.0"
+    "@scure/base": "^2.0.0",
+    "@noble/hashes": "^2.0.1"
   },
   "devDependencies": {
     "@linky/config": "workspace:*",
