+++ packages/linkstr/src/defaultRelays.ts
@@ -1,0 +1,5 @@
+export const DEFAULT_NOSTR_RELAYS = [
+  "wss://relay.damus.io",
+  "wss://nos.lol",
+  "wss://relay.0xchat.com",
+];
