+++ packages/linkstr/src/headless.ts
@@ -2,11 +2,13 @@ import { Effect } from "effect";
 import type { Layer } from "effect";
 import { linkstrServices } from "./composition";
 import type { LinkstrServices } from "./composition";
+import type { OutboxStore } from "./outbox/OutboxStore";
 import type { NostrSecretKey, RelayUrl } from "./domain/primitives";
 import { NostrTransportSimplePool } from "./services/NostrTransport";
 import type { NostrTransport } from "./services/NostrTransport";
 
 export interface LinkstrHeadlessConfig {
+  readonly outboxStore?: Layer.Layer<OutboxStore> | undefined;
   readonly secretKey: NostrSecretKey;
   readonly readRelays: ReadonlyArray<RelayUrl>;
   /** Read-only consumers (the service worker) omit this. */
@@ -29,6 +31,7 @@ export const runLinkstr = <A, E>(
       effect,
       linkstrServices({
         secretKey: config.secretKey,
+        outboxStore: config.outboxStore,
         readRelays: config.readRelays,
         writeRelays: config.writeRelays ?? [],
         transport: config.transport ?? NostrTransportSimplePool,
